server/requirements.txt
fastapi
uvicorn
raftify
python-multipart
python-dotenv
uuid
psycopg2-binary
SQLAlchemy
requests
pydantic